Adds asset classes that follow new APIs (#116)
# Description

This MR introduces a new set of asset interfaces. These interfaces
simplify the spawning of assets into the scene and initializing the
physics handle by hiding that into post-startup physics callbacks.
Essentially, users no longer need to worry about the `spawn()` and
`initialize()` calls themselves. The only interface they need to work
with are the configuration objects. This makes the code more flexible
and usable.

The assets use the following conventions for function names:
* `set_xxx()`: this implies that values are set into data buffers (but
not into the simulator)
* `write_xxx_to_sim()`: writes data into the simulator
* `update()`: update all data from the simulator

The above isn't an "ideal" function naming since there is no duality of
functions (a write has a read method, or a setter has a getter). But I
think they convey their functionality clearly from the naming. All data
is accessible only through the `data` class for the asset.

As a side note, the MR also renames all "dof" to "joint" to make it
consistent with the terminology in robotics.

## Type of change

<!-- As you go through the list, delete the ones that are not
applicable. -->

- New feature (non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file

---------

Signed-off-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>
Co-authored-by: Pascal Roth <57946385+pascal-roth@users.noreply.github.com>
Co-authored-by: jsmith-bdai <142246516+jsmith-bdai@users.noreply.github.com>