Corrects execution command in RL training tutorial docs (#311)
# Description

Currently, the command to play a trained RL model is lacking the flag
that will make it use the latest checkpoint, despite that behavior
being described as expected in the surrounding text. This MR adds the flag.

It also removes a statement that 'by default' the above command will
load the latest checkpoint, because it could be potentially confusing as
that behavior is only 'by default' when you use the flag.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there