Adds string checking to `orbit.sim.utils` function calls (#429)
# Description

Adds string checking to the functions which accept path-types, casting
them to a `str` if they are not already. Closing #421 in favor of this
fix, see discussion for initial rationale. This solution is superior
because this addresses all of the changes, not just the one causing a
test to fail.

Fixes #420 

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: AutonomousHansen <50837800+AutonomousHansen@users.noreply.github.com>