Fixes test scripts to report their status to the runner (#475)
# Description

@Mayankm96 recently found that tests failures weren't properly reporting
in runs of `orbit -t`. This is because we were previously relying on
individual test's main to exit with 1 if the test failed, but this
recently changed when we stopped them from exiting with `exit=False`
argument to `unittest.main()` to ensure the simulation app could be
closed properly.
 
We no longer call `simulation_app.close()` as this will cause the exit to not reach `run_all_tests.py`. The downside here is a bit more console spam.

Fixes #474

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: James Smith <142246516+jsmith-bdai@users.noreply.github.com>
Co-authored-by: Mayank Mittal <mittalma@leggedrobotics.com>
Co-authored-by: Mayank Mittal <12863862+Mayankm96@users.noreply.github.com>